included separate package.json for epub
