+++ epub3/nodejs_epub/replaceTagWithSymbol.js
