+++ epub3/package.json
@@ -1,0 +1,24 @@
+{
+  "name": "sicpxml2latex",
+  "version": "1.0.0",
+  "description": "convert xml to latex",
+  "main": "index.js",
+  "dependencies": {
+    "lz-string": "^1.4.4",
+    "xmldom": "^0.1.27",
+    "xpath": "0.0.27"
+  },
+  "devDependencies": {
+    "babel-cli": "^6.26.0",
+    "babel-preset-es2015": "^6.24.1",
+    "babel-preset-stage-2": "^6.24.1",
+    "babel-register": "^6.26.0"
+  },
+  "scripts": {
+    "format": "prettier --write ./nodejs_epub/*.js",
+    "start": "babel-node ./nodejs_epub/index.js",
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "Chan Ger Hean",
+  "license": "ISC"
+}
